RavenDB-18206 : refactoring - introduce AbstractRestoreBackupTask class, with 3 implementations : RestoreBackupTask, RestoreSnapshotTask, SingleShardRestoreBackupTask
