+++ src/Raven.Server/Documents/PeriodicBackup/Restore/IRestoreSource.cs
@@ -1,11 +1,12 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.IO;
 using System.IO.Compression;
 using System.Threading.Tasks;
 
 namespace Raven.Server.Documents.PeriodicBackup.Restore
 {
-    internal interface IRestoreSource
+    public interface IRestoreSource : IDisposable
     {
         Task<Stream> GetStream(string path);
