+++ src/Raven.Server/Documents/PeriodicBackup/Restore/RestoreFromS3.cs
@@ -33,7 +33,7 @@ namespace Raven.Server.Documents.PeriodicBackup.Restore
         public async Task<ZipArchive> GetZipArchiveForSnapshot(string path)
         {
             var blob = await _client.GetObjectAsync(path);
-            var file = await RestoreBackupTask.CopyRemoteStreamLocally(blob.Data, _tempPath);
+            var file = await AbstractRestoreBackupTask.CopyRemoteStreamLocally(blob.Data, _tempPath);
             return new DeleteOnCloseZipArchive(file, ZipArchiveMode.Read);
         }
 
@@ -61,10 +61,7 @@ namespace Raven.Server.Documents.PeriodicBackup.Restore
 
         public void Dispose()
         {
-            using (_client)
-            {
-                //base.Dispose();
-            }
+            _client?.Dispose();
         }
     }
 }
